refactor: move all files from `no-meteor` to node app package
