+++ imports/node-app/plugins/system-info/index.js
@@ -1,6 +1,6 @@
-import queries from "./queries";
-import resolvers from "./resolvers";
-import schemas from "./schemas";
+import queries from "./queries/index.js";
+import resolvers from "./resolvers/index.js";
+import schemas from "./schemas/index.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
