+++ imports/node-app/plugins/system-info/queries/systemInformation.js
