+++ imports/node-app/plugins/system-info/resolvers/Query/index.js
@@ -1,0 +1,5 @@
+import systemInformation from "./systemInformation.js";
+
+export default {
+  systemInformation
+};
