+++ imports/node-app/plugins/system-info/resolvers/index.js
@@ -1,0 +1,5 @@
+import Query from "./Query/index.js";
+
+export default {
+  Query
+};
