--- imports/plugins/core/system-info/server/no-meteor/queries/index.js
@@ -1,5 +1,0 @@
-import systemInformation from "./systemInformation";
-
-export default {
-  systemInformation
-};
