--- imports/plugins/core/system-info/server/no-meteor/resolvers/index.js
@@ -1,5 +1,0 @@
-import Query from "./Query";
-
-export default {
-  Query
-};
